+++ packages/beff-wasm/ts-node/bundler.ts
@@ -222,23 +222,7 @@ export class Bundler {
 }
 function serializeSettings(settings: BeffUserSettings) {
   return {
-    string_formats:
-      Object.fromEntries(
-        (settings.stringFormats ?? []).map((it) => [
-          it.name,
-          {
-            error_message: it.errorMessage,
-          },
-        ]),
-      ) ?? {},
-    number_formats:
-      Object.fromEntries(
-        (settings.numberFormats ?? []).map((it) => [
-          it.name,
-          {
-            error_message: it.errorMessage,
-          },
-        ]),
-      ) ?? {},
+    string_formats: (settings.stringFormats ?? []).map((it) => it.name) ?? [],
+    number_formats: (settings.numberFormats ?? []).map((it) => it.name) ?? [],
   };
 }