+++ packages/beff-wasm/ts-node/project.ts
@@ -1,6 +1,5 @@
 export type BeffCustomFormat = {
   name: string;
-  errorMessage?: string;
 };
 
 export type BeffUserSettings = {