+++ e2e-tests/openapi-schema/CHANGELOG.md
@@ -1,5 +1,13 @@
 # openapi-schema-generator
 
+## 1.0.97
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/client@0.0.133
+  - @beff/cli@0.0.133
+
 ## 1.0.96
 
 ### Patch Changes