+++ e2e-tests/standalone-parser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # node-server
 
+## 1.0.130
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/client@0.0.133
+  - @beff/cli@0.0.133
+
 ## 1.0.129
 
 ### Patch Changes